修复怪物生成位置错误的问题。 修复骨粒子实体不消失的问题。

+++ Assets/Scripts/Engine/Level/Level/LevelEngine_Behaviour.cs
@@ -42,9 +42,9 @@ namespace PVZEngine.Level
         {
             return spawnedID.ToArray();
         }
-        public int GetEnemySpawnX()
+        public float GetEnemySpawnX()
         {
-            return GetProperty<int>(EngineAreaProps.ENEMY_SPAWN_X);
+            return GetProperty<float>(EngineAreaProps.ENEMY_SPAWN_X);
         }
         public void GameOver(int type, Entity killer, string message)
         {
